+++ android/app/src/main/java/io/coursepick/coursepick/presentation/createcustomcourse/CreateCustomCourseViewModel.kt
@@ -85,7 +85,7 @@ class CreateCustomCourseViewModel
         fun handleExitAction() {
             if (segments.value.isEmpty()) {
                 viewModelScope.launch {
-                    _event.emit(CustomCourseUiEvent.Exit)
+                    _event.emit(CreateCustomCourseUiEvent.Exit)
                 }
             } else {
                 _showDiscardDialog.value = true
