+++ android/app/src/main/java/io/coursepick/coursepick/presentation/customcourse/CustomCoursesFragment.kt
@@ -7,6 +7,7 @@ import android.view.ViewGroup
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
 import io.coursepick.coursepick.databinding.FragmentCustomCoursesBinding
+import io.coursepick.coursepick.presentation.createcustomcourse.CoordinateUiModel
 import io.coursepick.coursepick.presentation.map.MapManager
 import io.coursepick.coursepick.presentation.createcustomcourse.CreateCustomCourseActivity
